+++ exports/OwcPinboard.types.ts
@@ -1,0 +1,1 @@
+export { OwcPinboardOptions } from '../src/pinboard/OwcPinboard.types.js';